Added Italicized, Underlined, Boldfaced to TokenType.
We set these if the italic, underline, or boldface attribute
is true in a dsData element.

This gives better results for e.g. Markdown, but offers less
control to the user and gives arguably worse results for
e.g. lisp.  (In a programming editor it can be useful to
have parentheses highlighted as bold, but in displayed text
one wants the opposite, I would think.)

Tests have not been adjusted.  I'm undecided about this
change.

See jgm/pandoc#3812.