HTML formatting: removed outer div.
This caused problems with spacing between paragraphs and
code blocks, since it disabled the margin collapsing that
would otherwise occur between the p and the pre, and gave
us twice as much space as desired.

A possible cost of this change is that the CSS rule
putting `overflow-x: auto` on the enclosing div is gone.
(Putting this on the pre makes the line numbers not appear.)