core: rename Skylighting to Skylighting.Core
This change renames the main Skylighting module in the core library so
that its name no longer clashes with the name we want to provide in the
'skylighting' package. In that package, a new 'Skylighting' module just
re-exports the Skylighting.Core and Skylighting.Syntax modules. This
behavior matches the export behavior prior to the package split.

This commit also moves the 'skylighting' program into the 'skylighting'
package so that it has access to the generated parser modules.