Use a pure Haskell regex implementation instead of pcre.
Unexported Regex.KDE implements KDE-compatible regular expressions in
pure Haskell. The implementation is not efficient, but it seems
good enough for this application, and it is desirable to avoid depending
on a C library.  (Available Haskell libraries weren't up to the task,
because they don't do back-references, captures, lookahead/behind.)

Benchmarks: with the Haskell regex implementation, benchmarks
are about 2x slower than with the C. Probably an acceptable
tradeoff.