Change Keyword constructor type [API change].
Keyword now takes keyword attributes and an Either value:
either Left a list name -- which then needs to be resolved
into a WordSet,
or Right a word set.

In tokenize, we modify the syntax map passed into config
by resolving all the Left values to Right values.

This is necessary because a list may contain an include element,
which includes other syntaxes in the syntax map.  The parser
doesn't have access to these, so this resolution has to be done
here.  Note, though, that this is somewhat inefficient; in
a document that parses 100 typescript snippets, for example,
this resolution will be done 100 times.  It would be better
to change the syntax map globally, but this can be put off
for a future optimization.

This fixes issues with keyword resolution in typescript (which includes
lists in javascript), and in other syntaxes as well, including
raku, tcsh, scala, fortran-free, and zsh.

Closes #124.