Move resolveKeywords from Tokenizer -> Parser.
Also, ensure that loadSyntaxesFromDir automatically applies
resolveKeywords.  This will ensure that the work is just done
once, and not every time we tokenize a snippet.
(The tokenizer will still apply resolveKeywords, so that
a Syntax that hasn't been resolved will still work, but the
run time cost of resolveKeywords is much lower if keywords
have been pre-resolved.)