Associate regex captures with a context.
This is necessary for proper implementation of the KDE
syntax highlighting algorithm.

Captures are context-specific.

"If dynamic is true, a rule can use placeholders representing the text
matched by a regular expression rule that switched to the current
context in its string or char attributes. In a string, the placeholder
%N (where N is a number) will be replaced with the corresponding capture
N from the calling regular expression, starting from 1. In a char the
placeholder must be a number N and it will be replaced with the first
character of the corresponding capture N from the calling regular
expression. ...  It is important to mention that a text captured in a
RegExpr rule is only stored for the switched context, specified in its
context attribute."
https://docs.kde.org/trunk5/en/kate/katepart/highlight.html#kate-highlight-default-styles

Closes #160.

Tokenizer changes:

- Add newtype for Captures.
- Modify ContextStack so it stores Captures with each Context.