Store compiled regexes in RE
This changes the RE type to (lazily) compile the regex when constructed,
rather than in the tokenizer. This allows us to avoid re-compiling
regexes for each separate tokenize call, instead sharing them globally.

We try to hide the internals of this, exposing the previous interfae
(RE { reString, reCaseSensitive }) with pattern synonyms.