Fix wordDetect to work around an issue in the new html.xml.
Update html syntax definition from upstream.

This uses WordDetect instead of Regex to match `<!DOCTYPE`.
But, officially, there is no word boundary between whitespace
and `<` because `<` is not a word character.  I'm really not
sure why it works in KDE, but the simplest way around this is
simply to disable the checking for a word boundary at the
beginning of the word in `wordDetect`. In practice, a check
at the end should suffice.