Use `-webkit-text-size-adjust` instead of unsetting inline-block (#202)
* Revert "Re-add `display: inline-block`, but only to print CSS."
  This reverts commit 15682b2aeeac3a0f28601a12f281afe4c7bfbce4.

* Revert "Remove display: inline-block from span CSS."
  This partially reverts commit
  0ec889227c4a603bc7f9262b7952cb7a660d1829

* Use `-webkit-text-size-adjust` instead of unsetting inline-block

Having `display: inline-block;` is useful. I frequently customize the
pandoc code block styles to allow for line highlights (see [1]).

The CSS styles for those line hightlights rely on the enclosing `span`
for a line being able to span the full width of the code block, which is
not possible without making the `span` be `display: inline-block;`.

The original issue reported in https://github.com/jgm/pandoc/issues/7248
was that specifically iOS Safari renders font sizes incorrectly.

This is a known iOS bug. I first discovered the workaround for it while
working with the tufte-css project (see [2]), and similar workarounds
are implemented in various CSS normalize/reset projects.

I have verified that including this `-webkit-text-size-adjust` property
fixes the original bug reported against pandoc. Given that the solution
in this PR allows keeping `display: inline-block;` which has its own
uses, fixes the original bug, and is an iOS-specific solution to an
iOS-specific problem, I believe that this is a more compelling long-term
solution (at least until iOS decides to change this behavior).

I, of course, defer to the maintainers of the skylighting project as to
whether you want this change. I've already worked around this problem in
my own projects that consume the output of pandoc- and
skylighting-highlighted code blocks (by basically vendoring these
changes into those projects), so I am not personally blocked. I'm
contributing this change upstream in the hopes that it might help
unblock others.

[1]: https://jez.io/pandoc-markdown-css-theme/features/#numbered-and-highlighted-lines
[2]: https://github.com/edwardtufte/tufte-css/issues/81#issuecomment-261953409